scripts/requirements.txt
pyyaml
requests
click
git+https://dev.hostsharing.net/r/ansible/hs.admin.api.git